requirements.txt
pytorch
numpy
opencv-python
torchvision
pillow
